requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
PyJWT==2.6.0
bcrypt==4.0.1
python-dotenv==1.0.0
